ДЗ-2

+++ delete.php
@@ -3,6 +3,12 @@
 include __DIR__ . '/autoload.php';
 
 $delete = \App\Models\Article::findById($_GET['id']);
-$delete->delete();
 
-header('Location: /admin.php');
+if (isset($delete)) {
+    $delete->delete();
+    header('Location: /admin.php');
+    exit;
+} else {
+    header('HTTP/1.1 404 Not Found');
+    exit;
+}
